// assets/games/tongits/script/proto/tongits.ts
export interface RoomInfo {
  roomId: number;
  roomName: string;
  /** 房间状态 */
  roomStatus: number;
  /** 最大座位数 */
  maxSeat: number;
}

/** 玩家在游戏中的完整信息 */
export interface TongitsPlayerInfo {
  /** 基础玩家信息 */
  playerInfo:
    | PlayerInfo
    | undefined;
  /** 手牌数量 (对其他玩家可见) */
  handCardCount: number;
  /** 是否是庄家 */
  isDealer: boolean;
  /** 已经打出的牌组 (明牌) */
  displayedMelds: Meld[];
  /** 自己的手牌 老 */
  handCards: number[];
  /** 是否可以发起挑战 */
  isFight: boolean;
  /** 时间戳 */
  countdown: number;
  /** 1:默认待选择 2:发起 3:接受 4:拒绝 5:烧死 */
  changeStatus: number;
  /** 1:不可操作, 2:select, 3:Action */
  status: number;
  /** 是否赢得比赛 */
  isWin: boolean;
  /** 点数 */
  cardPoint: number;
  /** 自动组牌 */
  isAuto: boolean;
  /** 自己的手牌 新 */
  groupCards: Cards[];
}

export interface TongitsBasePlayerInfo {
  /** 玩家ID */
  playerId: number;
  /** 1:默认待选择 2:发起 3:接受 4:拒绝 5:烧死 */
  changeStatus: number;
  /** 剩余时间 */
  countdown: number;
}

/** 牌组结构 */
export interface Meld {
  /** 组成牌组的牌 */
  cards: number[];
  /** 高亮的牌 (可见) */
  highlightCards: number;
  /** 这个牌组属于哪个玩家 */
  ownerId: number;
  /** 牌组的唯一ID (例如，玩家Melds数组的索引) */
  meldId: number;
  /** 牌组是否锁定 (动画表现) */
  locked: boolean;
}

/** 游戏状态的核心信息 */
export interface GameInfo {
  /** 当前操作玩家ID */
  actionPlayerId: number;
  /** 游戏阶段 1:初始 2:游戏中  3: 挑战中 4: 结算前  5:结算中 */
  status: number;
  /** 牌堆剩余牌数 */
  deckCardCount: number;
  /** 弃牌堆的牌 从下到上 */
  discardPile: number[];
  /** 底池信息 */
  pot:
    | PotInfo
    | undefined;
  /** 弃牌显示牌 */
  discardCard: number;
  /** 胜利类型  1 tongits 2 挑战 3 时间结束比大小: */
  winType: number;
  /** 视角玩家id */
  perspectiveId: number;
  /** 当前下注币值 */
  betAmount: number;
  /** 币值列表 */
  scores: number[];
}

/** 底池信息 */
export interface PotInfo {
  /** 底池金额 */
  base: number;
  /** 连续未被赢取的次数 */
  winCount: number;
  /** 当前所属玩家id */
  useId: number;
}
// ...
export interface JoinRoomRes {
  /** 房间信息 */
  roomInfo: RoomInfo | undefined;
  players: TongitsPlayerInfo[];
  /** 旁观者信息 */
  watchers: PlayerInfo[];
  /** 旁观者数量 */
  playersCount: number;
  /** 开麦用户列表 */
  speakers: PlayerInfo[];
  /** 自己信息 */
  self:
    | TongitsPlayerInfo
    | undefined;
  /** 游戏信息 */
  gameInfo:
    | GameInfo
    | undefined;
  /** 玩家设置 */
  playerSettings: PlayerSettings | undefined;
}

/**
 * 开始游戏
 * MessageType: TONGITS_ROOM_OWNER_START_GAME_REQ (3019)
 */
export interface StartGameReq {
}

/**
 * 开始游戏
 * MessageType: TONGITS_ROOM_OWNER_START_GAME_RES (3020)
 */
export interface StartGameRes {
}

/**
 * 游戏开始广播
 * MessageType: TONGITS_START_GAME_BROADCAST (3015)
 */
export interface GameStartBroadcast {
  /** 游戏初始状态 */
  gameInfo:
    | GameInfo
    | undefined;
  /** 座位上的玩家信息 */
  players: TongitsPlayerInfo[];
  userId: number;
}
// ...
export interface DrawCardBroadcast {
  playerId: number;
  userId: number;
  /** 当自动抽牌时  自己会有这个值 */
  drawnCard: number;
  /** 手牌数量 (对其他玩家可见) */
  handCardCount: number;
  /** 自己的手牌  自己会有这个值 */
  groupCards: Cards[];
}
// ...
export interface MeldCardBroadcast {
  playerId: number;
  /** 新打出的牌组 */
  newMeld:
    | Meld
    | undefined;
  /** 手牌数量 (对其他玩家可见) */
  handCardCount: number;
  userId: number;
}
// ...
export interface LayOffCardBroadcast {
  /** 操作者 */
  actionPlayerId: number;
  /** 补的牌 */
  cardAdded: number;
  /** 目标玩家 */
  targetPlayerId: number;
  /** 目标牌组 */
  targetMeldId: number;
  /** 手牌数量 (对其他玩家可见) */
  handCardCount: number;
  userId: number;
}
// ...
export interface DiscardCardBroadcast {
  playerId: number;
  discardedCard: number;
  /** 解锁的牌组ID列表 */
  unlockMelds: number[];
  /** 手牌数量 (对其他玩家可见) */
  handCardCount: number;
  /** 弃牌堆的牌 从下到上 */
  discardPile: number[];
  userId: number;
}
// ...
export interface TakeCardBroadcast {
  playerId: number;
  /** 吃牌后形成的新牌组 */
  newMeld:
    | Meld
    | undefined;
  /** 手牌数量 (对其他玩家可见) */
  handCardCount: number;
  /** 弃牌显示牌 */
  discard: number;
  userId: number;
}

/**
 * 挑战/pk
 * MessageType: TONGITS_CHALLENGE_ACTION_REQ (3013)
 */
export interface ChallengeReq {
  /** 2:发起 3:接受 4:拒绝 */
  changeStatus: number;
}

/**
 * 挑战/pk
 * MessageType: TONGITS_CHALLENGE_ACTION_RES (3014)
 */
export interface ChallengeRes {
  basePlayers: TongitsBasePlayerInfo[];
}

/**
 * 挑战广播
 * MessageType: TONGITS_CHALLENGE_BROADCAST (3026)
 */
export interface ChallengeBroadcast {
  playerId: number;
  basePlayers: TongitsBasePlayerInfo[];
  /** 点数 */
  cardPoint: number;
  userId: number;
}

/**
 * PK广播
 * MessageType: TONGITS_PK_BROADCAST (3027)
 */
export interface PKBroadcast {
  playerId: number;
  /** 2:发起 3:接受 4:拒绝 */
  changeStatus: number;
  userId: number;
}

/**
 * 游戏结算广播
 * MessageType: TONGITS_GAME_RESULT_BROADCAST (3016)
 */
export interface GameResultBroadcast {
  winnerId: number;
  playerResults: PlayerResult[];
  /** 时间戳 */
  countdown: number;
  userId: number;
}

/**
 * 房间重置广播
 * MessageType: TONGITS_ROOM_RESET_BROADCAST (3017)
 */
export interface RoomResetBroadcast {
  players: TongitsPlayerInfo[];
  /** 自己信息 */
  self:
    | TongitsPlayerInfo
    | undefined;
  /** 游戏信息 */
  gameInfo:
    | GameInfo
    | undefined;
  /** 通知玩家id */
  userId: number;
}

/**
 * 操作变动广播
 * MessageType: TONGITS_ACTION_CHANGE_BROADCAST (3018)
 */
export interface ActionChangeBroadcast {
  /** 操作玩家 */
  actionPlayerId: number;
  /** 倒计时 */
  countdown: number;
  /** 是否可以发起挑战 */
  isFight: boolean;
  /** 1:不可操作, 2:select, 3:Action */
  status: number;
  /** 自己的手牌 */
  groupCards: Cards[];
  /** 通知玩家id */
  userId: number;
}

/**
 * 结算前比牌 广播
 * MessageType: TONGITS_GAME_WIN_BROADCAST (3028)
 */
export interface BeforeResultBroadcast {
  winnerId: number;
  /** 胜利类型  1 tongits 2 挑战 3 时间结束比大小: */
  winType: number;
  players: TongitsPlayerInfo[];
  /** 时间戳 */
  countdown: number;
  /** 底池信息 */
  pot:
    | PotInfo
    | undefined;
  /** 通知玩家id */
  userId: number;
}

/**
 * 获取游戏结算详情数据
 * MessageType: TONGITS_GAME_RESULT_DETAILS_REQ (3029)
 */
export interface GameResultDetailsReq {
}

/**
 * 获取游戏结算详情数据
 * MessageType: TONGITS_GAME_RESULT_DETAILS_RES (3030)
 */
export interface GameResultDetailsRes {
  winnerId: number;
  playerResults: PlayerResult[];
}
// ...
export interface Cards {
  /** 牌组id   0是散牌  1 ... 是牌组 */
  groupId: number;
  /** 牌组手牌 */
  handCards: number[];
  /** 牌型 0 无效牌型 1 有效牌型 2特殊牌型 */
  cardType: number;
  /** 牌组点数 */
  cardPoint: number;
}
// ...
export const GameStartBroadcast = {
  encode(message: GameStartBroadcast, writer: _m0.Writer = _m0.Writer.create()): _m0.Writer {
    if (message.gameInfo !== undefined) {
      GameInfo.encode(message.gameInfo, writer.uint32(10).fork()).ldelim();
    }
    for (const v of message.players) {
      TongitsPlayerInfo.encode(v!, writer.uint32(18).fork()).ldelim();
    }
    if (message.userId !== 0) {
      writer.uint32(32).int64(message.userId);
    }
    return writer;
  },

  decode(input: _m0.Reader | Uint8Array, length?: number): GameStartBroadcast {
    const reader = input instanceof _m0.Reader ? input : _m0.Reader.create(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseGameStartBroadcast();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1:
          if (tag !== 10) {
            break;
          }

          message.gameInfo = GameInfo.decode(reader, reader.uint32());
          continue;
        case 2:
          if (tag !== 18) {
            break;
          }

          message.players.push(TongitsPlayerInfo.decode(reader, reader.uint32()));
          continue;
        case 4:
          if (tag !== 32) {
            break;
          }

          message.userId = longToNumber(reader.int64() as Long);
          continue;
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skipType(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): GameStartBroadcast {
    return {
      gameInfo: isSet(object.gameInfo) ? GameInfo.fromJSON(object.gameInfo) : undefined,
      players: globalThis.Array.isArray(object?.players)
        ? object.players.map((e: any) => TongitsPlayerInfo.fromJSON(e))
        : [],
      userId: isSet(object.userId) ? globalThis.Number(object.userId) : 0,
    };
  },

  toJSON(message: GameStartBroadcast): unknown {
    const obj: any = {};
    if (message.gameInfo !== undefined) {
      obj.gameInfo = GameInfo.toJSON(message.gameInfo);
    }
    if (message.players?.length) {
      obj.players = message.players.map((e) => TongitsPlayerInfo.toJSON(e));
    }
    if (message.userId !== 0) {
      obj.userId = Math.round(message.userId);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<GameStartBroadcast>, I>>(base?: I): GameStartBroadcast {
    return GameStartBroadcast.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<GameStartBroadcast>, I>>(object: I): GameStartBroadcast {
    const message = createBaseGameStartBroadcast();
    message.gameInfo = (object.gameInfo !== undefined && object.gameInfo !== null)
      ? GameInfo.fromPartial(object.gameInfo)
      : undefined;
    message.players = object.players?.map((e) => TongitsPlayerInfo.fromPartial(e)) || [];
    message.userId = object.userId ?? 0;
    return message;
  },
};
// ...
export const ActionChangeBroadcast = {
  encode(message: ActionChangeBroadcast, writer: _m0.Writer = _m0.Writer.create()): _m0.Writer {
    if (message.actionPlayerId !== 0) {
      writer.uint32(8).int64(message.actionPlayerId);
    }
    if (message.countdown !== 0) {
      writer.uint32(16).int32(message.countdown);
    }
    if (message.isFight !== false) {
      writer.uint32(24).bool(message.isFight);
    }
    if (message.status !== 0) {
      writer.uint32(32).int32(message.status);
    }
    for (const v of message.groupCards) {
      Cards.encode(v!, writer.uint32(42).fork()).ldelim();
    }
    if (message.userId !== 0) {
      writer.uint32(88).int64(message.userId);
    }
    return writer;
  },

  decode(input: _m0.Reader | Uint8Array, length?: number): ActionChangeBroadcast {
    const reader = input instanceof _m0.Reader ? input : _m0.Reader.create(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseActionChangeBroadcast();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1:
          if (tag !== 8) {
            break;
          }

          message.actionPlayerId = longToNumber(reader.int64() as Long);
          continue;
        case 2:
          if (tag !== 16) {
            break;
          }

          message.countdown = reader.int32();
          continue;
        case 3:
          if (tag !== 24) {
            break;
          }

          message.isFight = reader.bool();
          continue;
        case 4:
          if (tag !== 32) {
            break;
          }

          message.status = reader.int32();
          continue;
        case 5:
          if (tag !== 42) {
            break;
          }

          message.groupCards.push(Cards.decode(reader, reader.uint32()));
          continue;
        case 11:
          if (tag !== 88) {
            break;
          }

          message.userId = longToNumber(reader.int64() as Long);
          continue;
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skipType(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): ActionChangeBroadcast {
    return {
      actionPlayerId: isSet(object.actionPlayerId) ? globalThis.Number(object.actionPlayerId) : 0,
      countdown: isSet(object.countdown) ? globalThis.Number(object.countdown) : 0,
      isFight: isSet(object.isFight) ? globalThis.Boolean(object.isFight) : false,
      status: isSet(object.status) ? globalThis.Number(object.status) : 0,
      groupCards: globalThis.Array.isArray(object?.groupCards)
        ? object.groupCards.map((e: any) => Cards.fromJSON(e))
        : [],
      userId: isSet(object.userId) ? globalThis.Number(object.userId) : 0,
    };
  },

  toJSON(message: ActionChangeBroadcast): unknown {
    const obj: any = {};
    if (message.actionPlayerId !== 0) {
      obj.actionPlayerId = Math.round(message.actionPlayerId);
    }
    if (message.countdown !== 0) {
      obj.countdown = Math.round(message.countdown);
    }
    if (message.isFight !== false) {
      obj.isFight = message.isFight;
    }
    if (message.status !== 0) {
      obj.status = Math.round(message.status);
    }
    if (message.groupCards?.length) {
      obj.groupCards = message.groupCards.map((e) => Cards.toJSON(e));
    }
    if (message.userId !== 0) {
      obj.userId = Math.round(message.userId);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<ActionChangeBroadcast>, I>>(base?: I): ActionChangeBroadcast {
    return ActionChangeBroadcast.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<ActionChangeBroadcast>, I>>(object: I): ActionChangeBroadcast {
    const message = createBaseActionChangeBroadcast();
    message.actionPlayerId = object.actionPlayerId ?? 0;
    message.countdown = object.countdown ?? 0;
    message.isFight = object.isFight ?? false;
    message.status = object.status ?? 0;
    message.groupCards = object.groupCards?.map((e) => Cards.fromPartial(e)) || [];
    message.userId = object.userId ?? 0;
    return message;
  },
};
// ...
export type Exact<P, I extends P> = P extends Builtin ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };